[Refactor] 주문생성 리팩토링 (#252)
- 별도 서비스로 빼고, 함수로 조금 나누었습니다.
