+++ application-client/src/main/kotlin/org/fastcampus/applicationclient/order/controller/OrderController.kt
@@ -8,6 +8,7 @@ import org.fastcampus.applicationclient.order.controller.dto.response.OrderCreat
 import org.fastcampus.applicationclient.order.controller.dto.response.OrderDetailResponse
 import org.fastcampus.applicationclient.order.controller.dto.response.OrderResponse
 import org.fastcampus.applicationclient.order.service.OrderCancellationService
+import org.fastcampus.applicationclient.order.service.OrderCreationService
 import org.fastcampus.applicationclient.order.service.OrderService
 import org.fastcampus.common.dto.APIResponseDTO
 import org.fastcampus.common.dto.CursorDTO
@@ -28,6 +29,7 @@ import org.springframework.web.bind.annotation.RestController
 class OrderController(
     private val orderService: OrderService,
     private val orderCancellationService: OrderCancellationService,
+    private val orderCreationService: OrderCreationService,
 ) : OrderControllerDocs {
     @JwtAuthenticated
     @GetMapping
@@ -57,7 +59,7 @@ class OrderController(
         @RequestBody orderCreationRequest: OrderCreationRequest,
         @AuthenticationPrincipal authMember: AuthMember,
     ): APIResponseDTO<OrderCreationResponse> {
-        val response = orderService.createOrder(authMember.id, orderCreationRequest)
+        val response = orderCreationService.createOrder(authMember.id, orderCreationRequest)
         return APIResponseDTO(HttpStatus.OK.value(), HttpStatus.OK.reasonPhrase, response)
     }
