+++ application-client/src/test/kotlin/org/fastcampus/applicationclient/order/service/OrderCreationServiceTest.kt
@@ -40,7 +40,7 @@ class OrderCreationServiceTest {
     private lateinit var orderMenuRepository: OrderMenuRepository
     private lateinit var orderMenuOptionGroupRepository: OrderMenuOptionGroupRepository
     private lateinit var orderMenuOptionRepository: OrderMenuOptionRepository
-    private lateinit var orderService: OrderService
+    private lateinit var orderCreationService: OrderCreationService
 
     @BeforeEach
     fun init() {
@@ -51,7 +51,7 @@ class OrderCreationServiceTest {
         orderMenuRepository = mock(OrderMenuRepository::class.java)
         orderMenuOptionGroupRepository = mock(OrderMenuOptionGroupRepository::class.java)
         orderMenuOptionRepository = mock(OrderMenuOptionRepository::class.java)
-        orderService = OrderService(
+        orderCreationService = OrderCreationService(
             memberRepository = memberRepository,
             orderRepository = orderRepository,
             storeRepository = storeRepository,
@@ -100,7 +100,7 @@ class OrderCreationServiceTest {
             .thenAnswer { (it.arguments[0] as OrderMenuOption).copy(id = 1) }
 
         // when
-        val result = orderService.createOrder(1, request)
+        val result = orderCreationService.createOrder(1, request)
 
         // then
         expectThat(result) {
